+++ tests/unit/test_published_clusters.py
@@ -20,11 +20,13 @@ class PublishedClusterTest(TestCase):
         datasets_json = [self._published_clusters_json]
         project_id = "1"
 
-        project_url = f"http://localhost:9100/api/versioned/v1/projects/{project_id}"
-        unified_dataset_url = f"http://localhost:9100/api/versioned/v1/projects/{project_id}/unifiedDataset"
-        datasets_url = f"http://localhost:9100/api/versioned/v1/datasets"
-        refresh_url = f"http://localhost:9100/api/versioned/v1/projects/{project_id}/publishedClusters:refresh"
-        operations_url = f"http://localhost:9100/api/versioned/v1/operations/93"
+        project_url = f"{self._base_url}/projects/{project_id}"
+        unified_dataset_url = f"{self._base_url}/projects/{project_id}/unifiedDataset"
+        datasets_url = f"{self._base_url}/datasets"
+        refresh_url = (
+            f"{self._base_url}/projects/{project_id}/publishedClusters:refresh"
+        )
+        operations_url = f"{self._base_url}/operations/93"
 
         responses.add(responses.GET, project_url, json=self._project_config_json)
         responses.add(
@@ -44,7 +46,7 @@ class PublishedClusterTest(TestCase):
     @responses.activate
     def test_published_clusters_configuration(self):
         path = "projects/1/publishedClustersConfiguration"
-        config_url = f"http://localhost:9100/api/versioned/v1/{path}"
+        config_url = f"{self._base_url}/{path}"
         responses.add(responses.GET, config_url, json=self._config_json)
 
         p = Project(self.unify, self._project_config_json).as_mastering()
@@ -60,14 +62,25 @@ class PublishedClusterTest(TestCase):
 
     @responses.activate
     def test_refresh_ids(self):
-        path = "projects/1/allPublishedClusterIds:refresh"
-        refresh_url = f"http://localhost:9100/api/versioned/v1/{path}"
+        refresh_url = f"{self._base_url}/projects/1/allPublishedClusterIds:refresh"
         responses.add(responses.POST, refresh_url, json=self._operations_json)
 
         p = Project(self.unify, self._project_config_json).as_mastering()
         op = p.refresh_published_cluster_ids()
         self.assertEqual(op.resource_id, self._operations_json["id"])
         self.assertTrue(op.succeeded())
+
+    @responses.activate
+    def test_refresh_stats(self):
+        refresh_url = f"{self._base_url}/projects/1/publishedClusterStats:refresh"
+        responses.add(responses.POST, refresh_url, json=self._operations_json)
+
+        p = Project(self.unify, self._project_config_json).as_mastering()
+        op = p.refresh_published_cluster_stats()
+        self.assertEqual(op.resource_id, self._operations_json["id"])
+        self.assertTrue(op.succeeded())
+
+    _base_url = "http://localhost:9100/api/versioned/v1"
 
     _project_config_json = {
         "id": "unify://unified-data/v1/projects/1",